Adds ability to delete images.

+++ src/services/images/hooks/remove-from-disk.js
@@ -5,8 +5,12 @@ export default options => {
   return hook => {
     for (var i = 0; i < hook.result.length; i++) {
       const image = hook.result[i];
+      if(!hook.app.get('clientPath')) {
+        return new Error('Sorry client path is missing');
+      }
       const folderName = path.join(hook.app.get('clientPath'), image.url || 'al23gl239');
       fse.remove(folderName);
     }
+    return hook;
   }
 }
